third_party/WebKit/LayoutTests/external/wpt/html/webappapis/system-state-and-capabilities/the-navigator-object/historical.window-expected.txt

This is a testharness.js-based test.
PASS registerContentHandler() is removed
FAIL isProtocolHandlerRegistered() is removed assert_false: expected false got true
PASS isContentHandlerRegistered() is removed
PASS unregisterContentHandler() is removed
Harness: the test ran to completion.
